requirements.txt
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.174.0
beautifulsoup4==4.13.4
google-generativeai==0.8.3
python-dotenv==1.0.1
